get the repo dev ready (#34)
We have two namespaces on openshift, aos-art-web and art-build-dev. When the project on the dev branch is built on art-build-dev, we need the front end to connect to the backend server deployed on art-build-dev namespace. Similarly, when the project is built in the aos-art-web namespace, it needs to point to the server in that same namespace. We get the OPENSHIFT_BUILD_NAMESPACE inside the pod, hence we can use that to dynamically point to the required server.