Feature upgrades (#39)
* get the repo dev ready (#34)

We have two namespaces on openshift, aos-art-web and art-build-dev. When the project on the dev branch is built on art-build-dev, we need the front end to connect to the backend server deployed on art-build-dev namespace. Similarly, when the project is built in the aos-art-web namespace, it needs to point to the server in that same namespace. We get the OPENSHIFT_BUILD_NAMESPACE inside the pod, hence we can use that to dynamically point to the required server.

* set OPENSHIFT_BUILD_NAMESPACE in Dockerfile (#35)

Since nextjs is not reading the pod env variables, the next best solution is to define the variable in the Dockerfile and override the variable in build config.

* Update dockerfile, add jenkins build details (#36)

* set OPENSHIFT_BUILD_NAMESPACE in Dockerfile

Since nextjs is not reading the pod env variables, the next best solution is to define the variable in the Dockerfile and override the variable in build config.

* add jenkins build field

* enable exact search, add info icon (#37)

Enable partial search with quotes and add info icon with popover.

* partial search for Jenkins field, links to release controller (#38)

* partial search for jenkins field, release page fix

* add links to release controller

* update docs, add comments, remove unused code (#40)