[ART-5986] Allow URL filtering for build status page
Enhanced the BUILD_HISTORY_HOME component to support URL filtering. This update enables the users to filter the build status page based on various parameters directly through the URL. For instance, navigating to art-dash.engineering.redhat.com/dashboard/build/history?package=openshift&version=4.11 will automatically apply the respective filters on the page.

The changes also ensure that any filters applied on the page are reflected in the URL, which provides a more interactive and user-friendly experience.

This was achieved by:
- Importing and utilizing the useRouter hook from next/router to listen for route changes.
- Creating a function to update the URL with the applied filters.
- Utilizing useEffect hooks to initialize filter states from the URL and re-fetch data when search parameters change.